requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
black==23.10.1
isort==5.12.0
mypy==1.6.1
python-multipart==0.0.6